# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "LIS",
        "description": "Laboratory Information Service API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8089",
            "description": "Local server"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/lis-service/facility-laboratory-maps/{id}": {
            "get": {
                "tags": [
                    "Facility-laboratory maps"
                ],
                "summary": "Get a facility-to-lab-test mapping by id",
                "operationId": "getById_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityLaboratoryMapResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Facility-laboratory maps"
                ],
                "summary": "Update a facility-to-lab-test mapping",
                "operationId": "update_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFacilityLaboratoryMap"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityLaboratoryMapResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Facility-laboratory maps"
                ],
                "summary": "Delete a facility-to-lab-test mapping",
                "operationId": "delete_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/v1/lis-service/facility-laboratory-maps": {
            "get": {
                "tags": [
                    "Facility-laboratory maps"
                ],
                "summary": "List facility-to-lab-test mappings (paged), optionally filtered",
                "operationId": "list_4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    },
                    {
                        "name": "facilityId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "laboratoryTestId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "isActive",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedResponseFacilityLaboratoryMapResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Facility-laboratory maps"
                ],
                "summary": "Create a facility-to-lab-test mapping",
                "operationId": "create_4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFacilityLaboratoryMap"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FacilityLaboratoryMapResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lis-service/facility-laboratory-maps/bulk-upload": {
            "post": {
                "tags": [
                    "Facility-laboratory maps"
                ],
                "summary": "Bulk upload facility-to-lab-test mappings from a filled-in template",
                "description": "Existing active mappings are left unchanged, inactive ones are reactivated, and new combinations are created. Returns a summary of what happened.",
                "operationId": "bulkUpload",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkUploadSummary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/lis-service/reports/lab-orders": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Download the lab-order turn-around-time report (.xlsx)",
                "description": "One row per lab order for the facility, joined to its acknowledgement and current result. Tat is the order-to-result turn-around time (days + hours). Optionally bounded by an inclusive [from, to] date window on when the order was placed.",
                "operationId": "labOrderReport",
                "parameters": [
                    {
                        "name": "mflCode",
                        "in": "query",
                        "description": "Facility MFL code to scope the report to",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 101010
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start of the window (inclusive, ISO date). Omit for no lower bound.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End of the window (inclusive, ISO date). Omit for no upper bound.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                                "schema": {
                                    "type": "string",
                                    "format": "byte"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateFacilityLaboratoryMap": {
                "type": "object",
                "description": "Update a facility-to-lab-test mapping",
                "properties": {
                    "isActive": {
                        "type": "boolean"
                    }
                }
            },
            "FacilityLaboratoryMapResponse": {
                "type": "object",
                "description": "Facility-to-lab-test mapping",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "facilityId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "laboratoryTestId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CreateFacilityLaboratoryMap": {
                "type": "object",
                "description": "Map a facility to a lab-offered test",
                "properties": {
                    "facilityId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "laboratoryTestId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "facilityId",
                    "laboratoryTestId"
                ]
            },
            "BulkUploadSummary": {
                "type": "object",
                "description": "Summary of a facility-laboratory-map bulk upload",
                "properties": {
                    "totalRows": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total data rows read from the sheet (excludes the header and blank rows)",
                        "example": 120
                    },
                    "created": {
                        "type": "integer",
                        "format": "int32",
                        "description": "New mappings created",
                        "example": 80
                    },
                    "reactivated": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Existing inactive mappings reactivated",
                        "example": 5
                    },
                    "skippedExisting": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Existing active mappings left unchanged",
                        "example": 30
                    },
                    "duplicateRows": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Rows that duplicated an earlier row in the same file",
                        "example": 2
                    },
                    "errorRows": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Rows that could not be processed",
                        "example": 3
                    },
                    "errors": {
                        "type": "array",
                        "description": "Details of each row that failed",
                        "items": {
                            "$ref": "#/components/schemas/RowError"
                        }
                    }
                }
            },
            "RowError": {
                "type": "object",
                "description": "A single row that could not be processed",
                "properties": {
                    "row": {
                        "type": "integer",
                        "format": "int32",
                        "description": "1-based row number in the sheet (matches the spreadsheet)",
                        "example": 14
                    },
                    "mflCode": {
                        "type": "string"
                    },
                    "loincCode": {
                        "type": "string"
                    },
                    "labCode": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string",
                        "description": "Why the row was rejected",
                        "example": "Unknown mfl code"
                    }
                }
            },
...
            "PagedResponseFacilityLaboratoryMapResponse": {
                "type": "object",
                "description": "Paged response",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of items",
                        "example": "[{\"id\":1,\"name\":\"Item 1\"}, ...]",
                        "items": {
                            "$ref": "#/components/schemas/FacilityLaboratoryMapResponse"
                        }
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Total number of items",
                        "example": 100
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Total number of pages",
                        "example": 10
                    },
                    "currentPage": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Current page number",
                        "example": 1
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Number of items per page",
                        "example": 20
                    },
                    "hasNext": {
                        "type": "boolean"
                    },
                    "hasPrevious": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}